0.9.0 Los componentes se muestran en UI pero no estan actualizados ni ordenados

+++ app/src/main/java/com/example/testeo/Objects/Componente.java
@@ -18,7 +18,7 @@ public class Componente implements Serializable {
 
     }
 
-    public double getKm() {
+    public int getKm() {
         return km;
     }
